fix(go-rewrite): dedupe newGlobalStore test helper
PRs #423 and #425 each independently added the same helper in
package api_test, so post-merge `go test ./internal/api/...` fails
to build with `newGlobalStore redeclared`. Move the helper to a
single shared `helpers_external_test.go`.

Refs #407
