+++ server/go/internal/api/list_mailbox_users_test.go
@@ -16,19 +16,8 @@ import (
 
 	"github.com/elloloop/tenant-shard-db/server/go/internal/api"
 	"github.com/elloloop/tenant-shard-db/server/go/internal/errs"
-	"github.com/elloloop/tenant-shard-db/server/go/internal/globalstore"
 	pb "github.com/elloloop/tenant-shard-db/server/go/internal/pb"
 )
-
-func newGlobalStore(t *testing.T) *globalstore.GlobalStore {
-	t.Helper()
-	gs, err := globalstore.New(globalstore.Options{DataDir: t.TempDir(), WALMode: true})
-	if err != nil {
-		t.Fatalf("globalstore.New: %v", err)
-	}
-	t.Cleanup(func() { _ = gs.Close() })
-	return gs
-}
 
 // TestListMailboxUsers_EmptyForValidTenant pins the deprecated-stub
 // contract: a tenant that passes the gate returns an empty user_ids
